% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMI.R
\name{cMI}
\alias{cMI}
\title{Fast conditional Mutual Information}
\usage{
cMI(X, Y, Z, k = "suggest", alpha = 0.7, type = c("ps", "ksg"))
}
\description{
Computes the conditional mutual information between three random variables, i.e. I(X ; Y | Z)
Uses a fast range search to compute the neighbors in the max-norm. 
Relies on Kraskov's estimator
}
