% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMAC.R
\name{generateMAC}
\alias{generateMAC}
\title{Generates a MAC file}
\usage{
generateMAC(folder, number_to_generate_per_laminate_type = 500,
  delete_old = TRUE, uniform_variation = TRUE, n_ply = 24,
  E_fiber_nominal = 338200, E_fiber_percent_variation = 3,
  UTS_fiber_nominal = 3500, UTS_fiber_percent_variation = 5,
  nu_fiber_nominal = 0.3, nu_fiber_percent_variation = 3,
  E_matrix_nominal = 3450, E_matrix_percent_variation = 3,
  UTS_matrix_nominal = 80, UTS_matrix_percent_variation = 5,
  nu_matrix_nominal = 0.35, nu_matrix_percent_variation = 3,
  volume_fraction_nominal = 0.6, volume_fraction_variation = 0.03,
  angle_degree_variation = 1)
}
\arguments{
\item{constituents}{valid constituent data.frame.}

\item{analysis_type}{valid constituent data.frame.}

\item{loading}{valid constituent data.frame.}

\item{failure}{valid constituent data.frame.}
}
\description{
This function generates a MAC file for simulating tensile tests with MAC/GMC.
}
