% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStressStrainStats.R
\name{getStressStrainStats}
\alias{getStressStrainStats}
\title{getStressStrainStats}
\usage{
getStressStrainStats(strain, stress, uts_linearity_thresh = 0.98)
}
\description{
Attempts to detect various statistics along the stress-strain curve numerically.
}
