% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLayup.R
\name{makeLayup}
\alias{makeLayup}
\title{makeLayup}
\usage{
makeLayup(x)
}
\arguments{
\item{x}{layup orientation. See details.}
}
\description{
generates a ply layup using standard notation.
}
\examples{
makeLayup("[0]_24")
makeLayup("[0/90]_12")
makeLayup("[0, 15, 75, 90]_3S")
}
