% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN.R
\name{MLP}
\alias{MLP}
\title{MLP}
\usage{
MLP(layers)
}
\arguments{
\item{layers}{named-list of layer parameters, whose names correspond to layer names and value to layer parameters.}
}
\value{
a unary function right-folded using the magrittr pipe whose evaluation yields a keras model.
}
\description{
Creates a multilayer perceptron using \pkg{keras}'s functional API.
}
\details{
The \pkg{keras} package simplifies building neural networks in R using the 
\pkg{magrittr} pipe-compatible API to functionally compose layers. But individual layers 
are linked at the time of evaluation, making it cumbersome to e.g. programmatically change 
parameters at the individual layer level. This function uses the magrittr pipe ('%>%') 
to turn a simple named-list of layers + parameters into a functional sequence of activation layers.
}
\examples{

## Make simple regression model
X <- matrix(runif(100))
Y <- (0.5 * X + 2) + rnorm(100, sd=0.2)
layers <- list(
 "dense" = list(units = 1, activation = 'linear')
)

## Programmatically change the parameters 
for (i in seq(10)){
  layers[[1]]$units <- i
  model <- MLP(layers)
  ## ... 
}

}
