% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMAC.R
\name{parse_df}
\alias{parse_df}
\title{parse_df}
\usage{
parse_df(df, pre_space = "", collapse = FALSE)
}
\value{
if collapse=TRUE, a strign of the parsed rows is returned. Otherwise a list where each elements represents a row is returned.
}
\description{
Parses each row of a named data.frame into a tag list
}
