% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMAC.R
\name{parse_row}
\alias{parse_row}
\title{parse_row}
\usage{
parse_row(x, x_names = names(x), collapse = " ")
}
\value{
a character string containing a 'tag=value' for each column of a row.
}
\description{
Converts a vector with names into a tag=value space-delimited string
}
