% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseLists.R
\name{parseLists}
\alias{parseLists}
\title{Given a named list, recursively parses the names into sections, content into lines.}
\usage{
parseLists()
}
\description{
Given a named list, recursively parses the names into sections, content into lines.
}
