% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN.R
\name{partition}
\alias{partition}
\title{Partition}
\usage{
partition(X, Y, input_shape, splits = c(train = 0.8, validate = 0.1, test
  = 0.1))
}
\arguments{
\item{X}{data.table of input variables.}

\item{Y}{data.table of output variables.}
}
\value{
List with partitioned indices for the training, testing, and validation sets.
}
\description{
Partitions a data set into training, testing, and validation splits
}
