% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMAC.R
\name{runMAC}
\alias{runMAC}
\title{Runs MAC}
\usage{
runMAC(src_dir, dest_dir, base_fn, user, host = "cryoterm1",
  mac = "mac4z-3_8_TR", n_jobs = 25)
}
\description{
Runs MAC
}
\details{
Creates a batch file on the given host in the destination directory which, once executed, 
runs MAC/GMC on all of the input files in the destination directory. Assumes the destination directory 
contains the MAC/GMC executable.
}
