% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMAC.R
\name{validate_constituents}
\alias{validate_constituents}
\title{parse_constituents}
\usage{
validate_constituents(constituents)
}
\arguments{
\item{constituents}{data.frame of constituent materials. See details.}
}
\description{
Validates a data.frame of constituents for simulation with MAC/GMC. The data.frame is 
returned invisibly, and a preview of what the section would look like when produced with 
\code{\link{generateMAC}} is shown.
}
\details{
Must have at least 11 columns, plus optionally a name column.
}
\references{
MAC/GMC Reference Manual v4.0: https://ntrs.nasa.gov/archive/nasa/casi.ntrs.nasa.gov/20030015411.pdf
}
